requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
